Warning if pipe input has headers, etc.
Because raw PCM doesn't have any structure it can be hard to notice
that WAV data is accidentally going to the stdin pipe. This can cause
mysterious un-decodable signals or can shift the time-from-start
timestamps. We try to at least detect a WAV header in
supposedly-raw PCM data and give a warning to the user.